# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vo de Taxi",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/clients": {
            "get": {
                "tags": [
                    "#1 - Client"
                ],
                "summary": "Display all customers",
                "description": "Display all customers",
                "operationId": "display_clients",
                "responses": {
                    "201": {
                        "description": "Clients registered successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            },
            "post": {
                "tags": [
                    "#1 - Client"
                ],
                "summary": "Register a new customer",
                "description": "Register a new customer",
                "operationId": "register_client",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "city",
                                    "state",
                                    "address",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text"
                                    },
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "password"
                                    },
                                    "city": {
                                        "type": "text"
                                    },
                                    "state": {
                                        "type": "text"
                                    },
                                    "address": {
                                        "type": "text"
                                    },
                                    "phone": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": [
                                    {
                                        "name": "carlos",
                                        "email": "example@email.com.br",
                                        "password": "12345678",
                                        "city": "Suzano",
                                        "state": "SP",
                                        "address": "Rua 01",
                                        "phone": "1195252-8596",
                                        "account_validation": "0"
                                    }
                                ]
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/clients/activation/{id}": {
            "put": {
                "tags": [
                    "#1 - Client"
                ],
                "summary": "Activate customer account",
                "description": "Activate customer account",
                "operationId": "active_client",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Client id",
                        "required": true
                    },
                    {
                        "name": "account_validation",
                        "in": "query",
                        "description": "1 active and 0 deactive",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": [
                                    {
                                        "account_validation": "1"
                                    }
                                ]
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/clients/{id}": {
            "delete": {
                "tags": [
                    "#1 - Client"
                ],
                "summary": "Deactive customer account",
                "description": "Deactive customer account",
                "operationId": "destroy_client",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Client id",
                        "required": true
                    }
                ],
                "responses": {
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    }
}